requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
nltk>=3.8
spacy>=3.4.0
transformers>=4.20.0
torch>=1.12.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pydantic>=1.10.0
PyYAML>=6.0
joblib>=1.1.0